Created Proof of Concept Jupyter Notebook
In order to test, review and explain the processes used in this project
I've created a Jupyter Notebook. In this notebook I'll be showing how it
is possible to import and order the data sets. Once the concept works
then it can be abstracted out into a single executable file which does
all of the heavy lifting.