Rewrote Introduction & Part 1 of Notebook
The next stage is making the notebook easier to read and interpret for
new-comers to the repo. Thus I'm rewriting sections to make them flow
better, be easier to understand and explain my thoughts more.

Note that there is still the old content contained in the notebook.
I've done this intentially so I can use it to write the future sections.