Restructured the notebook to be clearer
I've spent a significant chunck of time refining the notebook so that it
is easier to follow for others reading it, as well as easier to use as
documentation when I develop the executable file.

The conclusion isn't completed yet and the appendix is still a bit messy
but for the most part the notebook is completed.