Created fetchStationList function
Although the framework was already in place, I fleshed out the content
of the fetchStationList function.

At the same time I added the weather station csv file to the root
directory. I might change this in the near future (especially once I
import other data types) but for now it will do.