Created first working version of script
This commit is a milestone. Now we have a script that automatically
collects data from BOM, extracts it, tidys it and centralises it in one
place.

Other interesting things to note in this commit is the inclusion of the
tidyUp function which removes the old zipped files, plus a new verbose
output of which station we are up to in order to track progress.