Created MaxTempWeatherStation class & refactored init
Beforehand, you would initialise each type of data uniquely. That was
because I assumed that the URL would be wildly different for different
bits of data. I was wrong: only small parts vary. As such I created the
function "_initialiseUniqueVariables" to cut down on repeated code.

I also created the MaxTempWeatherStation class which will collect the
data for the maximum temperature for a weather station.