Refactor main & added tidyStationDir
Firstly, I refactored the main() function so that I could more easily
separate it from the final execution. That way I can control the input
parameters really easily!

Next, I added a new funtion: tidyStationDir. This function makes it so
it is easier to identify the different datasets with human-readable
names.